% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_eigen_values.R
\name{adjusted_eigen_values}
\alias{adjusted_eigen_values}
\title{Regularized eigen-values}
\usage{
adjusted_eigen_values(
  X,
  method = c("Touloumis_equal", "Touloumis_unequal", "Strimmer", "pseudodet")
)
}
\arguments{
\item{X}{data matrix}

\item{method}{strink using either "Touloumis" or "Strimmer" method}
}
\description{
Regularized eigen-values for low rank matrix
}
