% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSubset.R
\name{bestSubsetSearch}
\alias{bestSubsetSearch}
\title{Evaluate information criteria for best subset selection}
\usage{
bestSubsetSearch(Y, data, variables, maxk = 5)
}
\arguments{
\item{data}{data.frame with columns corresponding to formula}

\item{variables}{array of variable names to be considered in the regression.  If variable should be considered as random effect, use '(1|A)'.}

\item{maxk}{evaluate all combinations of variables up to size maxk}

\item{exprObj}{matrix of expression data (g genes x n samples), or ExpressionSet, or EList returned by voom() from the limma package}
}
\description{
Evaluate information criteria for best subset selection
}
