% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\docType{class}
\name{eclairs-class}
\alias{eclairs-class}
\title{Class eclairs}
\description{
Class \code{eclairs}
}
\details{
Object storing:
\itemize{
 \item{U: }{orthonormal matrix with k columns representing the low rank component}
 \item{dSq: }{eigen-values so that \eqn{U diag(d^2) U^T} is the low rank component}
 \item{lambda: }{shrinkage parameter \eqn{\lambda} for the scaled diagonal component}
 \item{nu: }{diagonal value, \eqn{\nu}, of target matrix in shrinkage}
 \item{n: }{number of samples (i.e. rows) in the original data}
 \item{p: }{number of features (i.e. columns) in the original data}
 \item{k: }{rank of low rank component}
 \item{rownames: }{sample names from the original matrix}
 \item{colnames: }{features names from the original matrix}
 \item{method: }{method used for decomposition}
 \item{call: }{the function call}
}
}
