% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclairs.R
\name{eclairs}
\alias{eclairs}
\title{Estimate covariance/correlation with low rank and shrinkage}
\usage{
eclairs(
  X,
  k,
  lambda = NULL,
  compute = c("covariance", "correlation"),
  warmStart = NULL,
  p = nrow(X)
)
}
\arguments{
\item{X}{data matrix with n samples as rows and p features as columns}

\item{k}{the rank of the low rank component}

\item{lambda}{shrinkage parameter. If not specified, it is estimated from the data.}

\item{compute}{compute the 'covariance' (default) or 'correlation'}

\item{warmStart}{result of previous SVD to initialize values}

\item{p}{number of features of current or original data set following SVD.}
}
\value{
\link{eclairs} object storing:
\itemize{
 \item{U: }{orthonormal matrix with k columns representing the low rank component}
 \item{dSq: }{eigen-values so that \eqn{U diag(d^2) U^T} is the low rank component}
 \item{lambda: }{shrinkage parameter \eqn{\lambda} for the scaled diagonal component}
 \item{nu: }{diagonal value, \eqn{\nu}, of target matrix in shrinkage}
 \item{n: }{number of samples (i.e. rows) in the original data}
 \item{p: }{number of features (i.e. columns) in the original data}
 \item{k: }{rank of low rank component}
 \item{rownames: }{sample names from the original matrix}
 \item{colnames: }{features names from the original matrix}
 \item{method: }{method used for decomposition}
 \item{call: }{the function call}
}
}
\description{
Estimate the covariance/correlation between columns as the weighted sum of a low rank matrix and a scaled identity matrix.  The weight acts to shrink the sample correlation matrix towards the identity matrix or the sample covariance matrix towards a scaled identity matrix with constant variance.  An estimate of this form is useful because it is fast, and enables fast operations downstream.
}
\details{
Compute \eqn{U}, \eqn{d^2} to approximate the covariance/correlation matrix between columns of data matrix X by \eqn{U diag(d^2 (1-\lambda)) U^T + diag(\nu * \lambda)}.  When computing the covariance matrix \eqn{\nu} is the constant variance which is the mean of all feature-wise variances.  When computing the correlation matrix, \eqn{\nu = 1}.
}
