% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_shrink.R
\name{estimateMVN_EB}
\alias{estimateMVN_EB}
\title{Fit Multivariate Normal Inverse Wishart Model using Empirical Bayes}
\usage{
estimateMVN_EB(X, MAP = FALSE)
}
\arguments{
\item{X}{response matrix with responses on columns and samples on rows}

\item{MAP}{compute Maximum a posteriori estimate of covariance between responses}
}
\description{
Fit Multivariate Normal Inverse Wishart Model using Empirical Bayes
}
