% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCriterion.R
\name{mvBIC_fit}
\alias{mvBIC_fit}
\title{Evaluate multivariate BIC}
\usage{
mvBIC_fit(exprObj, formula, data, countLevels = TRUE, verbose = FALSE)
}
\arguments{
\item{exprObj}{matrix of expression data (g genes x n samples), or ExpressionSet, or EList returned by voom() from the limma package}

\item{formula}{specifies variables for the linear (mixed) model.  Must only specify covariates, since the rows of exprObj are automatically used as a response. e.g.: \code{~ a + b + (1|c)} Formulas with only fixed effects also work, and \code{lmFit()} followed by \code{contrasts.fit()} are run.}

\item{data}{data.frame with columns corresponding to formula}

\item{countLevels}{default TRUE, count number of levels rather than number of variance compinents.  See description in \code{\link{nparam}}}

\item{verbose}{Default TRUE. Print messages}
}
\value{
multivariate BIC value
}
\description{
Evaluate multivariate BIC while considering correlation between response variables. For n samples, p responses and m parameters for each model, evaluate the multivariate BIC as \deqn{n * logDet(\Sigma) + log(n) * (p*m + 0.5*p*(p+1))}
where \eqn{\Sigma} is the residual covariance matrix.  This formula extends the standard univariate BIC to the multivariate case.  
For one response the standard penalty is \eqn{log(n)*m}, this just adds a \eqn{log(n)} to that value, but only the differences between two models is important.  Estimating the \eqn{p x p} covariance matrix requires \eqn{0.5*p*(p+1)} parameters.
When \eqn{p > m} the residual covariance matrix Sigma is not full rank.  In this case the psudo-determinant is used instead.

See References

Pauler, DK. The Schwarz criterion and related methods for normal linear models. Biometrika (1998), 85, 1, pp. 13-27

Edward J. Bedrick and Chih-Ling Tsai. Model Selection for Multivariate Regression in Small Samples.  Biometrics,  50:1 1994 226-231

TJ Wu, P Chen, Y Yan. The weighted average information criterion for multivariate regression model selection. Signal Processing 93.1 (2013): 49-55.
}
\details{
Evaluate multivariate BIC on matrix of response variables.  Smaller is better.
}
\examples{

# create matrix of responses
Y = with(iris, rbind(Sepal.Width, Sepal.Length))

# Evaluate model 1
mvBIC_fit( Y, ~ Species, data=iris)

# Evaluate model 2
# smaller mvBIC means better model
mvBIC_fit( Y, ~ Petal.Width + Petal.Length + Species, data=iris)

}
