% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCriterion.R
\name{mvIC_fit}
\alias{mvIC_fit}
\title{Evaluate multivariate BIC}
\usage{
mvIC_fit(
  exprObj,
  formula,
  data,
  criterion = c("BIC", "sum BIC", "AIC", "AICC", "CAIC", "sum AIC"),
  shrink.method = c("EB", "none", "var_equal", "var_unequal"),
  nparamsMethod = c("edf", "countLevels", "lme4"),
  fastApprox = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{exprObj}{matrix of expression data (g genes x n samples), or ExpressionSet, or EList returned by voom() from the limma package}

\item{formula}{specifies variables for the linear (mixed) model.  Must only specify covariates, since the rows of exprObj are automatically used as a response. e.g.: \code{~ a + b + (1|c)} Formulas with only fixed effects also work, and \code{lmFit()} followed by \code{contrasts.fit()} are run.}

\item{data}{data.frame with columns corresponding to formula}

\item{criterion}{multivariate criterion ('AIC', 'BIC') or summing score assuming independence of reponses ('sum AIC', 'sum BIC')}

\item{shrink.method}{Shrink covariance estimates to be positive definite. Using "var_equal" assumes all variance on the diagonal are equal.  This method is the fastest because it is linear time.  Using "var_unequal" allows each response to have its own variance term, however this method is quadratic time.  Using "none" does not apply shrinkge, but is only valid when there are very few responses}

\item{nparamsMethod}{"edf": effective degrees of freedom. "countLevels" count number of levels in each random effect.  "lme4" number of variance compinents, as used by lme4.  See description in \code{\link{nparam}}}

\item{fastApprox}{use PCA to transform variables and substantially decrease compute time.  This approach is exact when only fixed effects are used, and approximate when random effects are included}

\item{verbose}{Default TRUE. Print messages}

\item{...}{additional arguements passed to logDet}
}
\value{
multivariate BIC value
}
\description{
Evaluate multivariate BIC while considering correlation between response variables. For n samples, p responses and m parameters for each model, evaluate the multivariate BIC as \deqn{n * logDet(\Sigma) + log(n) * (p*m + 0.5*p*(p+1))}
where \eqn{\Sigma} is the residual covariance matrix.  This formula extends the standard univariate BIC to the multivariate case.  
For one response the standard penalty is \eqn{log(n)*m}, this just adds a \eqn{log(n)} to that value, but only the differences between two models is important.  Estimating the \eqn{p x p} covariance matrix requires \eqn{0.5*p*(p+1)} parameters.
When \eqn{p > m} the residual covariance matrix Sigma is not full rank.  In this case the psudo-determinant is used instead.
}
\details{
Evaluate multivariate BIC on matrix of response variables.  Smaller is better.
}
\examples{

# create matrix of responses
Y = with(iris, rbind(Sepal.Width, Sepal.Length))

# Evaluate model 1
mvIC_fit( Y, ~ Species, data=iris)

# Evaluate model 2
# smaller mvIC means better model
mvIC_fit( Y, ~ Petal.Width + Petal.Length + Species, data=iris)

}
\references{
{
  \insertRef{pauler1998schwarz}{mvIC}

  \insertRef{bedrick1994model}{mvIC}

  \insertRef{wu2013weighted}{mvIC}
}
}
