% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCriterion.R
\name{mvIC_from_residuals}
\alias{mvIC_from_residuals}
\title{Evaluate multivariate BIC from matrix of residuals}
\usage{
mvIC_from_residuals(
  residMatrix,
  m,
  criterion = c("AIC", "BIC", "AICC", "CAIC", "sum AIC", "sum BIC"),
  shrink.method = c("var_equal", "var_unequal", "EB", "none"),
  ...
)
}
\arguments{
\item{residMatrix}{matrix of residuals where rows are features}

\item{m}{number of parameters for each model}

\item{criterion}{multivariate criterion ('AIC', 'BIC') or summing score assumign independence of reponses ('sum AIC', 'sum BIC')}

\item{shrink.method}{Shrink covariance estimates to be positive definite. Using "var_equal" assumes all variance on the diagonal are equal.  This method is the fastest because it is linear time.  Using "var_unequal" allows each response to have its own variance term, however this method is quadratic time.  Using "none" does not apply shrinkge, but is only valid when there are very few responses}

\item{...}{other arguments passed to logDet}
}
\description{
Evaluate multivariate BIC from matrix of residuals
}
