% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_eigen_values.R
\name{rlogDet}
\alias{rlogDet}
\title{Regularized log determinant}
\usage{
rlogDet(
  X,
  method = c("Touloumis_equal", "Touloumis_unequal", "Strimmer", "pseudodet"),
  ...
)
}
\arguments{
\item{X}{data matrix}

\item{method}{strink using either "Touloumis" or "Strimmer" method}
}
\description{
Regularized log determinant for low rank matrix
}
