% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSubset.R
\name{bestSubsetSearch}
\alias{bestSubsetSearch}
\title{Evaluate information criteria for best subset selection}
\usage{
bestSubsetSearch(Y, data, variables, maxk = 5)
}
\arguments{
\item{Y}{response matrix with responses ad columns and samples as rows}

\item{data}{data.frame with columns corresponding to formula}

\item{variables}{array of variable names to be considered in the regression.  Random effects are not allowed}

\item{maxk}{evaluate all combinations of variables up to size maxk}
}
\description{
Evaluate information criteria for best subset selection
}
