% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCriterion.R
\name{mvBIC_from_residuals}
\alias{mvBIC_from_residuals}
\title{Evaluate multivariate BIC from matrix of residuals}
\usage{
mvBIC_from_residuals(residMatrix, m)
}
\arguments{
\item{residMatrix}{matrix of residuals where rows are features}

\item{m}{number of parameters for each model}
}
\description{
Evaluate multivariate BIC from matrix of residuals
}
