% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCriterion.R
\name{mvBIC_from_residuals}
\alias{mvBIC_from_residuals}
\title{Evaluate multivariate BIC from matrix of residuals}
\usage{
mvBIC_from_residuals(
  residMatrix,
  m,
  useMVBIC = TRUE,
  logDetMethod = c("Touloumis_equal", "Touloumis_unequal", "Strimmer", "pseudodet"),
  ...
)
}
\arguments{
\item{residMatrix}{matrix of residuals where rows are features}

\item{m}{number of parameters for each model}

\item{useMVBIC}{(development only, do not change)}

\item{logDetMethod}{method to compute logDet of correlation matrix for finite sample size}
}
\description{
Evaluate multivariate BIC from matrix of residuals
}
