% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcTransform.R
\name{pcTransform}
\alias{pcTransform}
\title{PC transformation}
\usage{
pcTransform(X, varFrac = 0.95)
}
\arguments{
\item{X}{data matrix}

\item{varFrac}{fraction of variance to retain}
}
\description{
Transform data using PCA and retaining given fraction of variance
}
