% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_eigen_values.R
\name{rlogDet}
\alias{rlogDet}
\title{Regularized log determinant}
\usage{
rlogDet(X, shrink.method = c("var_equal", "var_unequal", "none"), ...)
}
\arguments{
\item{X}{data matrix}

\item{shrink.method}{Shrink covariance estimates to be positive definite. Using "var_equal" assumes all variance on the diagonal are equal.  This method is the fastest because it is linear time.  Using "var_unequal" allows each response to have its own variance term, however this method is quadratic time.  Using "none" does not apply shrinkge, but is only valid when there are very few responses}

\item{...}{additional arguments passed to adjusted_eigen_values}
}
\description{
Regularized log determinant for low rank matrix
}
